+++ src/pages/commits/ProjectSelect.tsx
@@ -5,7 +5,7 @@ import IconButton from "@leafygreen-ui/icon-button";
 import { uiColors } from "@leafygreen-ui/palette";
 import { Body, Label } from "@leafygreen-ui/typography";
 import { Input } from "antd";
-import Icon from "components/icons/Icon";
+import Icon from "components/Icons";
 import {
   GetProjectsQuery,
   GetProjectsQueryVariables,
