+++ src/pages/spawn/spawnHost/SpawnHostModal.tsx
@@ -5,7 +5,7 @@ import Button, { Variant } from "@leafygreen-ui/button";
 import { Subtitle } from "@leafygreen-ui/typography";
 import { AutoComplete, Input } from "antd";
 import { useSpawnAnalytics } from "analytics";
-import Icon from "components/icons/Icon";
+import Icon from "components/Icons";
 import { Modal } from "components/Modal";
 import { ModalContent, RegionSelector } from "components/Spawn";
 import { InputLabel } from "components/styles";
